requirements.txt
pytest==7.4.0
selenium==4.10.0
webdriver-manager==3.8.6
allure-pytest==2.13.2
pytest-xdist==3.3.1
allure-python-commons==2.13.2